metrics-collector/requirements.txt
psutil==6.1.0
python-logstash==0.4.8
